Remove redundant cast in BuiltinNumberFixed
The result of math.Pow() is already a float64, so there is no need to
re-cast it for further float64 operations with the (also float64)
underlying number value data.

+++ mellifera.go
@@ -7321,7 +7321,7 @@ func BuiltinNumberFixed(ctx *Context) *Builtin {
 		}
 
 		factor := math.Pow(10.0, float64(precision))
-		fixed := math.Round(delf.data*float64(factor)) / float64(factor)
+		fixed := math.Round(delf.data*factor) / float64(factor)
 		return ctx.NewNumber(fixed), nil
 	})
 }
